# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RentBase API
  version: 1.0.0
  description: |
    REST API for residential buildings, apartments and reservations.
    Swagger UI at /swagger is generated from the same controllers and stays in sync when the API is running.
    This file is the committed specification for every method.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/reservations:
    get:
      tags: [Reservations]
      summary: List own reservations, or all reservations for an administrator
      operationId: listReservations
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: userId
          schema: { type: integer }
          description: Administrator only.
      responses:
        '200':
          description: Reservation list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
    post:
      tags: [Reservations]
      summary: Create a reservation
      operationId: createReservation
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /api/reservations/{id}:
    put:
      tags: [Reservations]
      summary: Update a reservation
      operationId: updateReservation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
    delete:
      tags: [Reservations]
      summary: Cancel a reservation
      operationId: deleteReservation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Cancelled, no body.
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{id}/approval:
    patch:
      tags: [Users]
      summary: Approve or reject a registration
      operationId: setApproval
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '200':
          description: Approval updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{id}:
    delete:
      tags: [Users]
      summary: Delete a member
      operationId: deleteUser
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Deleted, no body.
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      in: path
      name: id
      required: true
      schema: { type: integer }
  responses:
    BadRequest:
      description: Malformed or incomplete payload.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unprocessable:
      description: Payload is well formed, but the business rule rejects it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    ApprovalRequest:
      type: object
      required: [isApproved]
      properties:
        isApproved: { type: boolean }
...
    UserResponse:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string }
        role: { type: string }
        isApproved: { type: boolean }
...
    ReservationRequest:
      type: object
      required: [apartmentId, startDate, endDate]
      properties:
        apartmentId: { type: integer }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        notes: { type: string }
    ReservationResponse:
      allOf:
        - $ref: '#/components/schemas/ReservationRequest'
        - type: object
          properties:
            id: { type: integer }
            apartmentNumber: { type: string }
            buildingName: { type: string }
            userId: { type: integer }
            userName: { type: string }
            createdAt: { type: string, format: date-time }